File type restriction, thumbnails

+++ gallery-app/src/main/java/lt/insoft/gallery/ui/viewmodel/GalleryVm.java
@@ -1,12 +1,16 @@
 package lt.insoft.gallery.ui.viewmodel;
 
+import static org.zkoss.zk.ui.util.Clients.alert;
+
 import java.io.Serializable;
 import java.util.List;
 
 import org.zkoss.bind.annotation.BindingParam;
 import org.zkoss.bind.annotation.Command;
 import org.zkoss.bind.annotation.Init;
+import org.zkoss.zk.ui.Executions;
 import org.zkoss.zk.ui.select.annotation.WireVariable;
+import org.zkoss.zul.Messagebox;
 
 import lombok.Getter;
 import lt.insoft.gallery.ui.helper.ImageViewHelper;
@@ -28,7 +32,15 @@ public class GalleryVm implements Serializable {
 
     @Command
     public void doSelectImage(@BindingParam("image") ImageThumbnail imageThumbnail) {
-        imageViewHelper.delete(imageThumbnail);
+        Messagebox.show("Are you sure you want to delete " + imageThumbnail.getName() + "?", "Question?", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION, event -> {
+            if (event.getName().equals("onOK")) {
+                imageViewHelper.delete(imageThumbnail);
+                alert("Image deleted!");
+                Executions.sendRedirect("/gallery");
+            } else {
+                alert("Deletion canceled!");
+            }
+        });
     }
 }
